# DAOS-5089 lru: Improve LRU/HASH implementation