--- a/src/cart/src/include/gurt/hash.h
+++ b/src/cart/src/include/gurt/hash.h
@@ -34,7 +34,7 @@
 #include <stdbool.h>
 
 #include <gurt/list.h>
-#include <gurt/types.h> /* for d_uuid */
+#include <gurt/types.h>
 
 /**
  * Hash table keeps and prints extra debugging information
@@ -56,67 +56,64 @@ extern "C" {
 
 typedef struct {
 	/**
-	 * Compare \p key with the key of the record \p rlink
+	 * Compare \p key with the key of the record \p link
 	 * This member function is mandatory.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The link chain of the record
-	 * \param[in] key	Key to compare
-	 * \param[in] ksize	Size of the key
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	The link chain of the record
+	 * \param[in]	key	Key to compare
+	 * \param[in]	ksize	Size of the key
 	 *
 	 * \retval	true	The key of the record equals to \p key.
 	 * \retval	false	No match
 	 */
-	bool	 (*hop_key_cmp)(struct d_hash_table *htable, d_list_t *rlink,
+	bool	 (*hop_key_cmp)(struct d_hash_table *htable, d_list_t *link,
 				const void *key, unsigned int ksize);
 	/**
-	 * Optional, generate a key for the record \p rlink.
+	 * Optional, generate a key for the record \p link.
 	 *
 	 * This function is called before inserting a record w/o key into a
 	 * hash table.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The link chain of the record to generate key.
-	 * \param[in] arg	Input arguments for the key generating.
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	The link chain of the record to generate key.
+	 * \param[in]	arg	Input arguments for the key generating.
 	 */
-	void	 (*hop_key_init)(struct d_hash_table *htable,
-				 d_list_t *rlink, void *arg);
-	/**
-	 * Optional, return the key of record \p rlink to \p key_pp, and size of
-	 * the key as the returned value.
-	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The link chain of the record being queried.
-	 * \param[out] key_pp	The returned key.
-	 *
-	 * \return		size of the key.
-	 */
-	int	 (*hop_key_get)(struct d_hash_table *htable, d_list_t *rlink,
-				void **key_pp);
+	void	 (*hop_key_init)(struct d_hash_table *htable, d_list_t *link,
+				 void *arg);
 	/**
 	 * Optional, hash \p key to a 32-bit value.
 	 * DJB2 hash is used when this function is abscent.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] key	Key to hash
-	 * \param[in] ksize	Key size
+	 * \param[in]	htable	hash table
+	 * \param[in]	key	Key to hash
+	 * \param[in]	ksize	Key size
 	 *
 	 * \return		hash of the key
 	 */
 	uint32_t (*hop_key_hash)(struct d_hash_table *htable, const void *key,
 				 unsigned int ksize);
 	/**
-	 * Optional, increase refcount on the record \p rlink
+	 * Get the hash of recorded key.
+	 *
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	the link to record.
+	 *
+	 * \return		hash of recorded key
+	 */
+	uint32_t (*hop_rec_hash)(struct d_hash_table *htable, d_list_t *link);
+
+	/**
+	 * Optional, increase refcount on the record \p link
 	 * If this function is provided, it will be called for successfully
 	 * inserted record.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The record being referenced.
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	The record being referenced.
 	 */
-	void	 (*hop_rec_addref)(struct d_hash_table *htable,
-				   d_list_t *rlink);
+	void	 (*hop_rec_addref)(struct d_hash_table *htable, d_list_t *link);
 	/**
-	 * Optional, release refcount on the record \p rlink
+	 * Optional, release refcount on the record \p link
 	 *
 	 * If this function is provided, it is called while deleting a record
 	 * from the hash table.
@@ -126,27 +123,26 @@ typedef struct {
 	 * If the record should not be automatically freed by the hash table
 	 * despite of refcount, then this function should never return true.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The rlink being released.
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	The link being released.
 	 *
 	 * \retval	false	Do nothing
 	 * \retval	true	Only if refcount is zero and the hash item
 	 *			can be freed. If this function can return
 	 *			true, then hop_rec_free() should be defined.
 	 */
-	bool	 (*hop_rec_decref)(struct d_hash_table *htable,
-				   d_list_t *rlink);
+	bool	 (*hop_rec_decref)(struct d_hash_table *htable, d_list_t *link);
 
 	/**
-	 * Optional, release multiple refcount on the record \p rlink
+	 * Optional, release multiple refcount on the record \p link
 	 *
 	 * This function expands on hop_rec_decref() so the notes from that
 	 * function apply here.  If hop_rec_decref() is not provided then
 	 * hop_rec_ndecref() shouldn't be either.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The rlink being released.
-	 * \param[in] count	The number of refs to be dropped.
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	The link being released.
+	 * \param[in]	count	The number of refcounts to be dropped.
 	 *
 	 * \retval	0	Do nothing
 	 * \retval	1	Only if refcount is zero and the hash item
@@ -155,17 +151,16 @@ typedef struct {
 	 *		negative value on error.
 	 */
 	int	 (*hop_rec_ndecref)(struct d_hash_table *htable,
-				    d_list_t *rlink,
-				    int count);
+				    d_list_t *link, int count);
 
 	/**
-	 * Optional, free the record \p rlink
+	 * Optional, free the record \p link
 	 * It is called if hop_decref() returns zero.
 	 *
-	 * \param[in] htable	hash table
-	 * \param[in] rlink	The record being freed.
+	 * \param[in]	htable	hash table
+	 * \param[in]	link	The record being freed.
 	 */
-	void	 (*hop_rec_free)(struct d_hash_table *htable, d_list_t *rlink);
+	void	 (*hop_rec_free)(struct d_hash_table *htable, d_list_t *link);
 } d_hash_table_ops_t;
 
 enum d_hash_feats {
@@ -210,6 +205,25 @@ enum d_hash_feats {
 	 * Note that if addref/decref are not provided this bit has no effect
 	 */
 	D_HASH_FT_EPHEMERAL	= (1 << 3),
+
+	/**
+	 * If the LRU bit is set:
+	 * The found in bucket item is moved on top of the list.
+	 * So, next search for it will be much faster.
+	 */
+	D_HASH_FT_LRU		= (1 << 4),
+
+	/**
+	 * Use Global Table Lock instead of per bucket locking.
+	 * TODO: should be removed when all will use per bucket locking.
+	 */
+	D_HASH_FT_GLOCK		= (1 << 15),
+};
+
+union d_hash_lock {
+	pthread_spinlock_t	spin;
+	pthread_mutex_t		mutex;
+	pthread_rwlock_t	rwlock;
 };
 
 struct d_hash_bucket {
@@ -221,13 +235,9 @@ struct d_hash_bucket {
 
 struct d_hash_table {
 	/** different type of locks based on ht_feats */
-	union {
-		pthread_spinlock_t	ht_spin;
-		pthread_mutex_t		ht_mutex;
-		pthread_rwlock_t	ht_rwlock;
-	};
+	union d_hash_lock	 ht_lock;
 	/** bits to generate number of buckets */
-	unsigned int		 ht_bits;
+	uint32_t		 ht_bits;
 	/** feature bits */
 	uint32_t		 ht_feats;
 #if D_HASH_DEBUG
@@ -244,8 +254,9 @@ struct d_hash_table {
 	d_hash_table_ops_t	*ht_ops;
 	/** array of buckets */
 	struct d_hash_bucket	*ht_buckets;
+	/** different type of locks based on ht_feats */
+	union d_hash_lock	*ht_locks;
 };
-
 
 /**
  * Create a new hash table.
@@ -261,8 +272,8 @@ struct d_hash_table {
  *
  * \return			0 on success, negative value on error
  */
-int  d_hash_table_create(uint32_t feats, unsigned int bits,
-			 void *priv, d_hash_table_ops_t *hops,
+int  d_hash_table_create(uint32_t feats, uint32_t bits, void *priv,
+			 d_hash_table_ops_t *hops,
 			 struct d_hash_table **htable_pp);
 
 /**
@@ -281,11 +292,11 @@ int  d_hash_table_create(uint32_t feats, unsigned int bits,
  *
  * \return			0 on success, negative value on error
  */
-int  d_hash_table_create_inplace(uint32_t feats, unsigned int bits,
-				 void *priv, d_hash_table_ops_t *hops,
+int  d_hash_table_create_inplace(uint32_t feats, uint32_t bits, void *priv,
+				 d_hash_table_ops_t *hops,
 				 struct d_hash_table *htable);
 
-typedef int (*d_hash_traverse_cb_t)(d_list_t *rlink, void *arg);
+typedef int (*d_hash_traverse_cb_t)(d_list_t *link, void *arg);
 
 /**
  * Traverse a hash table, call the traverse callback function on every item.
@@ -299,8 +310,8 @@ typedef int (*d_hash_traverse_cb_t)(d_list_t *rlink, void *arg);
  *
  * \return			zero on success, negative value if error.
  */
-int d_hash_table_traverse(struct d_hash_table *htable,
-			  d_hash_traverse_cb_t cb, void *arg);
+int  d_hash_table_traverse(struct d_hash_table *htable,
+			   d_hash_traverse_cb_t cb, void *arg);
 
 /**
  * Destroy a hash table.
@@ -335,36 +346,36 @@ int  d_hash_table_destroy(struct d_hash_table *htable, bool force);
 int  d_hash_table_destroy_inplace(struct d_hash_table *htable, bool force);
 
 /**
- * lookup \p key in the hash table, the found chain rlink is returned on
+ * lookup \p key in the hash table, the found chain link is returned on
  * success.
  *
  * \param[in] htable		Pointer to the hash table
  * \param[in] key		The key to search
  * \param[in] ksize		Size of the key
  *
- * \return			found chain rlink
+ * \return			found chain link
  */
 d_list_t *d_hash_rec_find(struct d_hash_table *htable, const void *key,
 			  unsigned int ksize);
 
 /**
  * Lookup \p key in the hash table, if there is a matched record, it should be
- * returned, otherwise \p rlink will be inserted into the hash table. In the
- * later case, the returned link chain is the input \p rlink.
+ * returned, otherwise \p link will be inserted into the hash table. In the
+ * later case, the returned link chain is the input \p link.
  *
  * \param[in] htable		Pointer to the hash table
  * \param[in] key		The key to be inserted
  * \param[in] ksize		Size of the key
- * \param[in] rlink		The link chain of the record being inserted
+ * \param[in] link		The link chain of the record being inserted
  *
  * \return			matched record
  */
 d_list_t *d_hash_rec_find_insert(struct d_hash_table *htable,
 				 const void *key, unsigned int ksize,
-				 d_list_t *rlink);
+				 d_list_t *link);
 
 /**
- * Insert a new key and its record chain \p rlink into the hash table. The hash
+ * Insert a new key and its record chain \p link into the hash table. The hash
  * table holds a refcount on the successfully inserted record, it releases the
  * refcount while deleting the record.
  *
@@ -374,28 +385,28 @@ d_list_t *d_hash_rec_find_insert(struct d_hash_table *htable,
  * \param[in] htable		Pointer to the hash table
  * \param[in] key		The key to be inserted
  * \param[in] ksize		Size of the key
- * \param[in] rlink		The link chain of the record being inserted
+ * \param[in] link		The link chain of the record being inserted
  * \param[in] exclusive		The key has to be unique if it is true.
  *
  * \return			0 on success, negative value on error
  */
 int  d_hash_rec_insert(struct d_hash_table *htable, const void *key,
-		       unsigned int ksize, d_list_t *rlink,
+		       unsigned int ksize, d_list_t *link,
 		       bool exclusive);
 
 /**
  * Insert an anonymous record (w/o key) into the hash table.
- * This function calls hop_key_init() to generate a key for the new rlink
+ * This function calls hop_key_init() to generate a key for the new link
  * under the protection of the hash table lock.
  *
  * \param[in] htable		Pointer to the hash table
- * \param[in] rlink		The link chain of the hash record
+ * \param[in] link		The link chain of the hash record
  * \param[in] arg		Arguments for key generating
  *
  * \return			0 on success, negative value on error
  */
-int  d_hash_rec_insert_anonym(struct d_hash_table *htable, d_list_t *rlink,
-			       void *arg);
+int  d_hash_rec_insert_anonym(struct d_hash_table *htable, d_list_t *link,
+			      void *arg);
 
 /**
  * Delete the record identified by \p key from the hash table.
@@ -411,26 +422,50 @@ bool d_hash_rec_delete(struct d_hash_table *htable, const void *key,
 		       unsigned int ksize);
 
 /**
- * Delete the record linked by the chain \p rlink.
+ * Delete the record linked by the chain \p link.
  * This record will be freed if hop_rec_free() is defined and the hash table
  * holds the last refcount.
  *
  * \param[in] htable		Pointer to the hash table
- * \param[in] rlink		The link chain of the record
+ * \param[in] link		The link chain of the record
  *
  * \retval			true	Successfully deleted the record
  * \retval			false	The record has already been unlinked
  *					from the hash table
  */
-bool d_hash_rec_delete_at(struct d_hash_table *htable, d_list_t *rlink);
+bool d_hash_rec_delete_at(struct d_hash_table *htable, d_list_t *link);
+
+/**
+ * Evict the record identified by \p key from the hash table.
+ *
+ * \param[in] htable		Pointer to the hash table
+ * \param[in] key		The key of the record being evicted
+ * \param[in] ksize		Size of the key
+ *
+ * \retval			true	Item with \p key has been evicted
+ * \retval			false	Can't find the record by \p key
+ */
+bool d_hash_rec_evict(struct d_hash_table *htable, const void *key,
+		      unsigned int ksize);
+
+/**
+ * Evict the record linked by the chain \p link.
+ *
+ * \param[in] htable		Pointer to the hash table
+ * \param[in] link		The link chain of the record
+ *
+ * \retval			true	Item has been evicted
+ * \retval			false	Not LRU feature
+ */
+bool d_hash_rec_evict_at(struct d_hash_table *htable, d_list_t *link);
 
 /**
  * Increase the refcount of the record.
  *
  * \param[in] htable		Pointer to the hash table
- * \param[in] rlink		The link chain of the record
+ * \param[in] link		The link chain of the record
  */
-void d_hash_rec_addref(struct d_hash_table *htable, d_list_t *rlink);
+void d_hash_rec_addref(struct d_hash_table *htable, d_list_t *link);
 
 /**
  * Decrease the refcount of the record.
@@ -438,9 +473,9 @@ void d_hash_rec_addref(struct d_hash_table *htable, d_list_t *rlink);
  * is set.
  *
  * \param[in] htable		Pointer to the hash table
- * \param[in] rlink		Chain rlink of the hash record
+ * \param[in] link		Chain link of the hash record
  */
-void d_hash_rec_decref(struct d_hash_table *htable, d_list_t *rlink);
+void d_hash_rec_decref(struct d_hash_table *htable, d_list_t *link);
 
 /**
  * Decrease the refcount of the record by count.
@@ -448,28 +483,27 @@ void d_hash_rec_decref(struct d_hash_table *htable, d_list_t *rlink);
  *
  * \param[in] htable		Pointer to the hash table
  * \param[in] count		Number of references to drop
- * \param[in] rlink		Chain rlink of the hash record
+ * \param[in] link		Chain link of the hash record
  *
  * \retval			0		Success
  * \retval			-DER_INVAL	Not enough references were held.
  */
-int d_hash_rec_ndecref(struct d_hash_table *htable, int count,
-		       d_list_t *rlink);
+int  d_hash_rec_ndecref(struct d_hash_table *htable, int count, d_list_t *link);
 
 /**
  * Check if the link chain has already been unlinked from the hash table.
  *
- * \param[in] rlink		The link chain of the record
+ * \param[in] link		The link chain of the record
  *
  * \retval			true	Yes
  * \retval			false	No
  */
-bool d_hash_rec_unlinked(d_list_t *rlink);
+bool d_hash_rec_unlinked(d_list_t *link);
 
 /**
- * Return the first entry in a hash table.  Do this by traversing the table, and
- * returning the first rlink value provided to the callback.
- * Returns rlink on success, or NULL on error or if the hash table is empty.
+ * Return the first entry in a hash table.  Do this by traversing the table,
+ * and returning the first link value provided to the callback.
+ * Returns link on success, or NULL on error or if the hash table is empty.
  *
  * Note this does not take a reference on the returned entry and has no ordering
  * semantics.  It's main use is for draining a hash table before calling
@@ -477,7 +511,7 @@ bool d_hash_rec_unlinked(d_list_t *rlink);
  *
  * \param[in] htable		Pointer to the hash table
  *
- * \retval			rlink	Pointer to first element in hash table
+ * \retval			link	Pointer to first element in hash table
  * \retval			NULL	Hash table is empty or error occurred
  */
 d_list_t *d_hash_rec_first(struct d_hash_table *htable);
@@ -508,20 +542,20 @@ void d_hash_table_debug(struct d_hash_table *htable);
  * set bit 0 to 1.
  */
 enum {
-	D_HTYPE_PTR		= 0, /**< pointer type handle */
+	D_HTYPE_PTR		= 0,	/**< pointer type handle */
 	/* Must enlarge D_HTYPE_BITS to add more types */
 };
 
 struct d_hlink;
 struct d_hlink_ops {
 	/** free callback */
-	void	(*hop_free)(struct d_hlink *rlink);
+	void	(*hop_free)(struct d_hlink *hlink);
 };
 
 struct d_rlink {
 	d_list_t		rl_link;
-	unsigned int		rl_ref;
-	unsigned int		rl_initialized:1;
+	uint32_t		rl_ref:30,
+				rl_initialized:1;
 };
 
 struct d_hlink {
@@ -530,11 +564,11 @@ struct d_hlink {
 	struct d_hlink_ops	*hl_ops;
 };
 
-struct d_hhash;
+struct d_hhash;			/**< internal definition */
 
-int  d_hhash_create(uint32_t feats, unsigned int bits, struct d_hhash **hhash);
-void d_hhash_destroy(struct d_hhash *hh);
-void d_hhash_hlink_init(struct d_hlink *hlink, struct d_hlink_ops *ops);
+int  d_hhash_create(uint32_t feats, uint32_t bits, struct d_hhash **hhash);
+void d_hhash_destroy(struct d_hhash *hhash);
+void d_hhash_hlink_init(struct d_hlink *hlink, struct d_hlink_ops *hl_ops);
 /**
  * Insert to handle hash table.
  * If \a type is D_HTYPE_PTR, user MUST ensure the bit 0 of \a hlink pointer is
@@ -543,7 +577,7 @@ void d_hhash_hlink_init(struct d_hlink *hlink, struct d_hlink_ops *ops);
  * type.
  */
 void d_hhash_link_insert(struct d_hhash *hhash, struct d_hlink *hlink,
-		         int type);
+			 int type);
 struct d_hlink *d_hhash_link_lookup(struct d_hhash *hhash, uint64_t key);
 void d_hhash_link_getref(struct d_hhash *hhash, struct d_hlink *hlink);
 void d_hhash_link_putref(struct d_hhash *hhash, struct d_hlink *hlink);
@@ -552,8 +586,8 @@ bool d_hhash_link_empty(struct d_hlink *hlink);
 void d_hhash_link_key(struct d_hlink *hlink, uint64_t *key);
 int  d_hhash_key_type(uint64_t key);
 bool d_hhash_key_isptr(uint64_t key);
-int  d_hhash_set_ptrtype(struct d_hhash *hhtab);
-bool d_hhash_is_ptrtype(struct d_hhash *hhtab);
+int  d_hhash_set_ptrtype(struct d_hhash *hhash);
+bool d_hhash_is_ptrtype(struct d_hhash *hhash);
 
 /******************************************************************************
  * UUID Hash Table Wrapper
@@ -577,23 +611,21 @@ struct d_ulink_ops {
 struct d_ulink {
 	struct d_rlink		 ul_link;
 	struct d_uuid		 ul_uuid;
-	/** optional argument for compare callback */
-	void			*ul_cmp_args;
 	struct d_ulink_ops	*ul_ops;
 };
 
-int  d_uhash_create(uint32_t feats, unsigned int bits,
-		    struct d_hash_table **uhtab);
-void d_uhash_destroy(struct d_hash_table *uhtab);
-void d_uhash_ulink_init(struct d_ulink *ulink, struct d_ulink_ops *rl_ops);
+int  d_uhash_create(uint32_t feats, uint32_t bits,
+		    struct d_hash_table **htable);
+void d_uhash_destroy(struct d_hash_table *htable);
+void d_uhash_ulink_init(struct d_ulink *ulink, struct d_ulink_ops *ul_ops);
 bool d_uhash_link_empty(struct d_ulink *ulink);
 bool d_uhash_link_last_ref(struct d_ulink *ulink);
-void d_uhash_link_addref(struct d_hash_table *uhtab, struct d_ulink *hlink);
-void d_uhash_link_putref(struct d_hash_table *uhtab, struct d_ulink *hlink);
-void d_uhash_link_delete(struct d_hash_table *uhtab, struct d_ulink *hlink);
-int  d_uhash_link_insert(struct d_hash_table *uhtab, struct d_uuid *key,
-			 void *cmp_args, struct d_ulink *hlink);
-struct d_ulink *d_uhash_link_lookup(struct d_hash_table *uhtab,
+void d_uhash_link_addref(struct d_hash_table *htable, struct d_ulink *ulink);
+void d_uhash_link_putref(struct d_hash_table *htable, struct d_ulink *ulink);
+void d_uhash_link_delete(struct d_hash_table *htable, struct d_ulink *ulink);
+int  d_uhash_link_insert(struct d_hash_table *htable, struct d_uuid *key,
+			 void *cmp_args, struct d_ulink *ulink);
+struct d_ulink *d_uhash_link_lookup(struct d_hash_table *htable,
 				    struct d_uuid *key, void *cmp_args);
 
 #if defined(__cplusplus)