… src/cart/src/utest/test_gurt.c
… src/cart/src/utest/test_linkage.cpp
… src/client/dfuse/dfuse_core.c
… src/common/lru.c
… src/common/misc.c
… src/common/tests/lru.c
… src/container/srv_target.c
… src/include/daos/lru.h
… src/placement/pl_map.c
… src/pool/srv_target.c
… src/rdb/rdb_kvs.c
… src/vos/vos_obj.c
… src/vos/vos_obj.h
… src/vos/vos_obj_cache.c
… utils/node_local_test.py