The Hash Table implementation was redesigned to make the locking
per bucket. It led to a change in the API. The new function
hop_rec_hash was introduced.

uint32_t (*hop_rec_hash)(struct d_hash_table *htable, d_list_t *link);

It should return the hash value of key which was used to store it
into the hash table.

New LRU feature was add into Hash Table implementation. The idea of
this feature is to popup the last found item into head of per bucket
list. This cause next search for this item much faster.

Basing on this LRU feature the DAOS LRU code was changed to make it
more scalable to maintain many hot items.

Performance testing of new implementations show the following results:

New Hash Table implementation.

Parallel execution of the same operations into several threads:

current master version (the time in processor tiks):
Operation (count)  | min time | avg time | max time
------------ | ---------- | --------- | -----------
insert(122880) | 200,407,811  | 301,821,520  |  390,762,741
traverse(122880) | 26,457,490 | 142,478,282 | 251,781,644
lookup(122880) | 3,091,640,839 | 3,180,550,249 | 3,231,423,491
delete(122880) | 289,866,694 | 320,578,773 | 330,955,234

this patch version (the time in processor tiks):
Operation (count)  | min time | avg time | max time
------------ | ---------- | --------- | -----------
insert(122880) |  20,992,422 |  33,196,314 | 55,259,780
traverse(122880) |  25,937,342 |  30,285,950 | 31,452,028
lookup(122880) | 324,108,436 | 325,161,848 | 325,536,972
delete(122880) | 23,429,403 | 28,689,445 | 30,819,230

The performance improvements is about 10 times.

Concurrent execution of the different operations in parallel into
several threads:

current master | this patch
---------------- | -----------
1,864,564,804 | 237,032,281

The performance improvements again is about 10 times.

New LRU implementation.

The test just store and the release a random values into LRU with
different size.

lru_test <size of LRU in 2^n> <count of elements>

$ time ./lru_test 4 100000000

time | current master | this patch
----- | ---------------- | -----------
real | 0m17.729s | 0m20.490s
user | 0m17.570s | 0m20.299s
sys | 0m0.059s | 0m0.091s

$ time ./lru_test 6 100000000

time | current master | this patch
----- | ---------------- | -----------
real | 0m21.342s | 0m19.487s
user | 0m21.192s | 0m19.336s
sys | 0m0.048s | 0m0.050s

$ time ./lru_test 8 100000000

time | current master | this patch
----- | ---------------- | -----------
real | 0m22.481s | 0m21.606s
user | 0m22.270s | 0m21.445s
sys | 0m0.109s | 0m0.060s

$ time ./lru_test 12 100000000

time | current master | this patch
----- | ---------------- | -----------
real | 0m26.401s | 0m23.433s
user | 0m26.235s | 0m23.274s
sys | 0m0.064s | 0m0.057s

$ time ./lru_test 16 100000000

time | current master | this patch
----- | ---------------- | -----------
real | 0m51.480s | 0m48.181s
user | 0m51.309s | 0m48.019s
sys | 0m0.062s | 0m0.053s

The new implementation has better scaling with LRU size.