@daosbuild1 2020-06-08
Test stage checkpatch completed with status FAILURE.  https://build.hpdd.intel.com//job/daos-stack/job/daos/view/change-requests/job/PR-2864/8/execution/node/59/log