@dmiter 2020-06-09
The tests failed because of hdf5 plugin has to be modified according API change. This should be done after this patch lands.