@dmiter 2020-06-09
@mchaarawi As I wrote in description: "The Hash Table implementation was redesigned to make the locking per bucket. It led to a change in the API." This API change is required to make possible a fine grane locking (per backet). In parallel access this will bring 10x performance improvements. So, I hope this change is make sence.