@jolivier23 2020-06-09
> @mchaarawi As I wrote in description: "The Hash Table implementation was redesigned to make the locking per bucket. It led to a change in the API." This API change is required to make possible a fine grane locking (per backet). In parallel access this will bring 10x performance improvements. So, I hope this change is make sence.

can the patch be made in such a was as to be backwards compatible?   One option would be to create a new API and internally use the new one while leaving the old one in place.   A 10x improvement in performance is nice but it's a bit of a pain if it breaks a bunch of stuff downstream.

Since it is a breaking change, normally, this would entail a major version bump on libdaos

By the way, it would be visually easier to see differences if you added commas in your cycles per operation numbers above.   It's hard to tell a difference between 10 and 9 digit numbers otherwise.