@daosbuild1 2020-06-11
Test stage Scan CentOS 7 RPMs completed with status FAILURE.  https://build.hpdd.intel.com/job/daos-stack/job/daos/job/PR-2864/13/display/redirect