@dmiter 2020-06-11
The failed tests use old  /lib64/libgurt.so.4 version instead of new one  /lib64/libgurt.so.5.