@ashleypittman 2020-07-27
> This patch was rebased on top of master from 2020-07-25

If that's the case then there's a significant performance regression, I'm looking at the mdtest results from these two links:

https://build.hpdd.intel.com/job/daos-stack/job/daos/job/PR-2864/38/artifact/Functional/job-2020-07-26T15.27-3e19a83-io-mdtest_small/job.log/*view*/
https://build.hpdd.intel.com/job/daos-stack/job/daos/job/master/2583/artifact/Functional/job-2020-07-27T01.33-50e22cb-io-mdtest_small/job.log/*view*/

Master:
`2020-07-27 01:36:07,881 process          L0479 DEBUG| [stdout] mdtest-3.3.0+dev was launched with 8 total task(s) on 2 node(s)
2020-07-27 01:36:07,881 process          L0479 DEBUG| [stdout] Command line used: mdtest '-a' 'POSIX' '-b' '2' '-z' '5' '-I' '10' '-i' '1' '-e' '4096' '-d' '/tmp/daos_dfuse' '-w' '4096'
2020-07-27 01:36:07,881 process          L0479 DEBUG| [stdout] Path: /tmp
2020-07-27 01:36:07,881 process          L0479 DEBUG| [stdout] FS: 36.5 GiB   Used FS: 9.9%   Inodes: 2.3 Mi   Used Inodes: 4.2%
2020-07-27 01:36:07,881 process          L0479 DEBUG| [stdout] 
2020-07-27 01:36:07,881 process          L0479 DEBUG| [stdout] Nodemap: 10101010
2020-07-27 01:36:07,882 process          L0479 DEBUG| [stdout] 8 tasks, 5040 files/directories
2020-07-27 01:36:35,671 process          L0479 DEBUG| [stdout] 
2020-07-27 01:36:35,671 process          L0479 DEBUG| [stdout] SUMMARY rate: (of 1 iterations)
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    Operation                      Max            Min           Mean        Std Dev
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    ---------                      ---            ---           ----        -------
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    Directory creation        :       2950.137       2950.000       2950.069          0.067
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    Directory stat            :       3005.902       3005.890       3005.896          0.005
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    Directory removal         :        647.334        647.321        647.328          0.005
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    File creation             :       1422.057       1422.022       1422.039          0.017
2020-07-27 01:36:35,672 process          L0479 DEBUG| [stdout]    File stat                 :       1215.884       1215.880       1215.882          0.002
2020-07-27 01:36:35,673 process          L0479 DEBUG| [stdout]    File read                 :       1186.637       1186.636       1186.637          0.000
2020-07-27 01:36:35,673 process          L0479 DEBUG| [stdout]    File removal              :       1213.387       1213.386       1213.386          0.000
2020-07-27 01:36:35,673 process          L0479 DEBUG| [stdout]    Tree creation             :        529.142        529.142        529.142          0.000
2020-07-27 01:36:35,673 process          L0479 DEBUG| [stdout]    Tree removal              :        156.634        156.634        156.634          0.000
2020-07-27 01:36:35,673 process          L0479 DEBUG| [stdout] -- finished at 07/27/2020 01:36:35 --
`

PR:

`2020-07-26 15:42:10,834 process          L0479 DEBUG| [stdout] mdtest-3.3.0+dev was launched with 8 total task(s) on 2 node(s)
2020-07-26 15:42:10,835 process          L0479 DEBUG| [stdout] Command line used: mdtest '-a' 'POSIX' '-b' '2' '-z' '5' '-I' '10' '-i' '1' '-e' '4096' '-d' '/tmp/daos_dfuse' '-w' '4096'
2020-07-26 15:42:10,835 process          L0479 DEBUG| [stdout] Path: /tmp
2020-07-26 15:42:10,835 process          L0479 DEBUG| [stdout] FS: 36.5 GiB   Used FS: 9.9%   Inodes: 2.3 Mi   Used Inodes: 4.2%
2020-07-26 15:42:10,835 process          L0479 DEBUG| [stdout] 
2020-07-26 15:42:10,835 process          L0479 DEBUG| [stdout] Nodemap: 10101010
2020-07-26 15:42:10,835 process          L0479 DEBUG| [stdout] 8 tasks, 5040 files/directories
2020-07-26 15:46:08,854 process          L0479 DEBUG| [stdout] 
2020-07-26 15:46:08,855 process          L0479 DEBUG| [stdout] SUMMARY rate: (of 1 iterations)
2020-07-26 15:46:08,855 process          L0479 DEBUG| [stdout]    Operation                      Max            Min           Mean        Std Dev
2020-07-26 15:46:08,855 process          L0479 DEBUG| [stdout]    ---------                      ---            ---           ----        -------
2020-07-26 15:46:08,855 process          L0479 DEBUG| [stdout]    Directory creation        :        179.128        179.128        179.128          0.000
2020-07-26 15:46:08,855 process          L0479 DEBUG| [stdout]    Directory stat            :        167.739        167.739        167.739          0.000
2020-07-26 15:46:08,855 process          L0479 DEBUG| [stdout]    Directory removal         :        142.206        142.205        142.206          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout]    File creation             :        126.155        126.155        126.155          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout]    File stat                 :        147.573        147.573        147.573          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout]    File read                 :        132.114        132.114        132.114          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout]    File removal              :        166.435        166.434        166.434          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout]    Tree creation             :         83.462         83.462         83.462          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout]    Tree removal              :         57.706         57.706         57.706          0.000
2020-07-26 15:46:08,856 process          L0479 DEBUG| [stdout] -- finished at 07/26/2020 15:46:08 --
`

Simply by looking at the wallclock time for this it's increased from 28 seconds to 238 seconds.